Separate LLM module
