+++ app/lessons/models.py
@@ -7,24 +7,19 @@ from wagtail.fields import RichTextField
 from wagtail.admin.panels import FieldPanel, InlinePanel
 from modelcluster.fields import ParentalKey
 from modelcluster.models import ClusterableModel
-import os
-from openai import OpenAI
 import logging
-from pydantic import BaseModel, Field, ValidationError
-from typing import List
 from django_htmx.http import HttpResponseClientRedirect
 from django.contrib.auth import get_user_model
 
+from .llm import NO_KEY_CONCEPT, get_llm_response, render_llm_prompt
+
 # Set up logging
 logger = logging.getLogger(__name__)
-
-LLM_MODEL = "gpt-4o-2024-08-06"
 
 # Define the character limit constant
 MAX_USER_MESSAGE_LENGTH = 100
 
 # Define the constant for responses without a key concept
-NO_KEY_CONCEPT = "NO_KEY_CONCEPT"
 SUCCESS_PARAM = "success"
 START_OVER_PARAM = "start_over"
 
@@ -71,37 +66,6 @@ class TranscriptMessage(models.Model):
 
     def __str__(self):
         return f"{self.role} message in {self.transcript}"
-
-
-class Suggestion(BaseModel):
-    """
-    Represents a suggestion for the user to consider in response to a chat message.
-
-    Attributes:
-        text (str): The text of the suggestion.
-    """
-
-    text: str
-
-
-class ChatResponse(BaseModel):
-    """
-    Represents a response from a chat assistant, including the assistant's message,
-    a list of suggestions, and the key concept addressed in the response.
-
-    Attributes:
-        assistant_message (str): The message generated by the chat assistant.
-        suggestions (List[Suggestion]): A list of suggestions related to the chat response.
-        addressed_key_concept (str): The single key concept addressed in this response,
-            or 'NO_KEY_CONCEPT' if none was used.
-    """
-
-    assistant_message: str
-    suggestions: List[Suggestion]
-    addressed_key_concept: str = Field(
-        ...,
-        description=f"The single key concept addressed in this response, or '{NO_KEY_CONCEPT}' if none was used",
-    )
 
 
 class KeyConcept(models.Model):
@@ -185,24 +149,6 @@ class Lesson(Page, ClusterableModel):
     ]
 
     def get_context(self, request: HttpRequest) -> dict:
-        """
-        Generate and return the context dictionary for the lesson page.
-
-        This method extends the default context with additional information
-        specific to the lesson, such as key concepts and a language model (LLM)
-        prompt.
-
-        :param request: The HTTP request object.
-        :return: A dictionary containing the context for rendering the lesson page.
-
-        The context includes:
-        - key_concepts: A queryset of all key concepts associated with the lesson.
-        - max_message_length: The maximum allowed length for user messages.
-        - llm_prompt: A rendered string template that serves as a prompt for the
-            language model, incorporating lesson-specific details such as difficulty
-            level, location, system prompt, key concepts, conversation history, and
-            language.
-        """
         context = super().get_context(request)
         context["key_concepts"] = self.key_concepts.all()
         context["max_message_length"] = MAX_USER_MESSAGE_LENGTH
@@ -216,9 +162,7 @@ class Lesson(Page, ClusterableModel):
             "conversation_history": request.session.get("conversation_history", []),
             "language": self.language,
         }
-        context["llm_prompt"] = render_to_string(
-            "lessons/prompt_template.txt", prompt_context
-        )
+        context["llm_prompt"] = render_llm_prompt(prompt_context)
 
         # Get or create a transcript
         context["transcript"] = self.get_or_create_transcript(request)
@@ -226,19 +170,6 @@ class Lesson(Page, ClusterableModel):
         return context
 
     def serve(self, request: HttpRequest) -> HttpResponse:
-        """
-        Serve the lesson based on the user's request method and parameters.
-
-        Handles GET requests with SUCCESS_PARAM, checks if all key concepts are responded to,
-        resets lesson progress if needed, and redirects accordingly. This is to ensure that the
-        success page is only shown when the lesson is complete and to correctly render the lesson
-        page if the user hasn't responded to all key concepts but somehow reached the success page.
-
-        Handles POST requests by validating user message length, updating responded key concepts,
-        getting a response from the LLM model, and checking lesson completion.
-
-        Returns appropriate HTTP responses based on the request method and lesson progress.
-        """
         if request.method == "GET" and SUCCESS_PARAM in request.GET:
             if self.user_has_responded_to_all_key_concepts(request):
                 self.reset_lesson_progress(request)
@@ -405,130 +336,7 @@ class Lesson(Page, ClusterableModel):
         return set(lesson_key_concepts) == set(responded_key_concepts)
 
     def get_llm_response(self, request: HttpRequest, user_message: str) -> HttpResponse:
-        """
-        Generates a response from a language model based on the user's message and the lesson context.
-
-        This method constructs a prompt using the lesson's context and the user's message, sends it to
-        the language model, and processes the response. It updates the conversation history and addressed
-        key concepts in the session, and returns an HTTP response with the assistant's message and suggestions.
-
-        :param request: The HTTP request object containing session data.
-        :param user_message: The message input by the user.
-        :return: An HTTP response containing the assistant's message and suggestions.
-
-        :raises ValidationError: If there is an error in validating the response data.
-        :raises Exception: For any unexpected errors during the process.
-        """
-
-        conversation_history = request.session.get("conversation_history", [])
-
-        addressed_key_concepts = request.session.get("addressed_key_concepts", [])
-        prompt = self.get_context(request)["llm_prompt"]
-        messages = (
-            [
-                {"role": "system", "content": prompt},
-            ]
-            + conversation_history
-            + [{"role": "user", "content": user_message}]
-        )
-
-        try:
-            client = OpenAI(api_key=os.environ.get("OPENAI_API_KEY"))
-            completion = client.beta.chat.completions.parse(
-                model=LLM_MODEL,
-                messages=messages,
-                response_format=ChatResponse,
-            )
-
-            chat_response = completion.choices[0].message
-
-            if chat_response.refusal:
-                logger.warning(f"OpenAI API refusal: {chat_response.refusal}")
-                return HttpResponse(
-                    render_to_string(
-                        "lessons/combined_htmx_response.html",
-                        {
-                            "error": "I'm sorry, but I can't respond to that request. Please try a different question or topic."
-                        },
-                    )
-                )
-
-            response_data = chat_response.parsed
-
-            # Get valid key concepts
-            valid_key_concepts = [
-                concept.concept for concept in self.key_concepts.all()
-            ]
-
-            # Validate key concept
-            if response_data.addressed_key_concept not in valid_key_concepts:
-                logger.warning(
-                    f"Invalid key concept: {response_data.addressed_key_concept}"
-                )
-                response_data.addressed_key_concept = NO_KEY_CONCEPT
-            elif (
-                response_data.addressed_key_concept != NO_KEY_CONCEPT
-                and response_data.addressed_key_concept not in addressed_key_concepts
-            ):
-                addressed_key_concepts.append(response_data.addressed_key_concept)
-
-            # Update conversation history
-            conversation_history.append({"role": "user", "content": user_message})
-            conversation_history.append(
-                {"role": "assistant", "content": response_data.assistant_message}
-            )
-            # Limit to last 10 messages
-            conversation_history = conversation_history[-10:]
-
-            # Update session variables
-            request.session["conversation_history"] = conversation_history
-            request.session["addressed_key_concepts"] = addressed_key_concepts
-
-            # Log assistant message
-            self.log_message(
-                request,
-                "assistant",
-                response_data.assistant_message,
-                response_data.addressed_key_concept,
-                LLM_MODEL,
-            )
-
-            # Render the combined response
-            combined_response = render_to_string(
-                "lessons/combined_htmx_response.html",
-                {
-                    "page": self,
-                    "assistant_message": response_data.assistant_message,
-                    "suggestions": response_data.suggestions,
-                    "addressed_key_concept": response_data.addressed_key_concept,
-                    "addressed_key_concepts": addressed_key_concepts,
-                    "valid_key_concepts": valid_key_concepts,
-                    "no_key_concept": NO_KEY_CONCEPT,
-                },
-            )
-
-            return HttpResponse(combined_response)
-
-        except ValidationError as e:
-            logger.error(f"Validation error in get_llm_response: {str(e)}")
-            return HttpResponse(
-                render_to_string(
-                    "lessons/combined_htmx_response.html",
-                    {
-                        "error": "An error occurred while processing the response. Please try again."
-                    },
-                )
-            )
-        except Exception as e:
-            logger.error(f"Unexpected error in get_llm_response: {str(e)}")
-            return HttpResponse(
-                render_to_string(
-                    "lessons/combined_htmx_response.html",
-                    {
-                        "error": "An unexpected error occurred. Please try again later or contact support if the problem persists."
-                    },
-                )
-            )
+        return get_llm_response(request, user_message, self)
 
     def log_message(self, request, role, content, key_concept=None, llm_model=None):
         transcript_id = request.session.get("transcript_id")
